+++ contracts/smart-account-interfaces/src/auth/policy/interface.rs
@@ -1,0 +1,10 @@
+use soroban_sdk::{auth::Context, contractclient, Address, Env, Vec};
+
+#[contractclient(name = "SmartAccountPolicyClient")]
+pub trait SmartAccountPolicy {
+    fn on_add(env: &Env, source: Address);
+    fn on_revoke(env: &Env, source: Address);
+    fn is_authorized(env: &Env, source: Address, contexts: Vec<Context>) -> bool;
+}
+
+
