+++ contracts/smart-account-interfaces/src/auth/policy/mod.rs
@@ -1,0 +1,1 @@
+pub mod interface;
