+++ contracts/smart-account-interfaces/src/lib.rs
@@ -1,12 +1,14 @@
 #![no_std]
 
 pub mod account;
+pub mod auth;
+pub mod error;
 pub mod plugin;
-pub mod policy;
 
 pub use account::{SmartAccountClient, SmartAccountInterface};
-pub use plugin::{SmartAccountPlugin, SmartAccountPluginClient};
-pub use policy::{
+pub use auth::policy::interface::{SmartAccountPolicy, SmartAccountPolicyClient};
+pub use auth::types::{
     Ed25519Signer, ExternalPolicy, Secp256r1Signer, Signer, SignerKey, SignerPolicy, SignerRole,
-    SmartAccountError, SmartAccountPolicy, SmartAccountPolicyClient,
 };
+pub use error::SmartAccountError;
+pub use plugin::{SmartAccountPlugin, SmartAccountPluginClient};
